refactor: organize transaction route tests by workflow
